[ovsp4rt] Implement 24-bit VNI values (#663)
- Modified `EncodeVniValue()` to accept a 32-bit input and encode
  the low-order 24 bits.

- Modified `EncodeTunnelId()` to accept a 32-bit input and encode
  the low-order 20 bits.

- Widened `tunnel_info.vni` from uint16_t to uint32_t.

- Enabled 24-bit VNI test cases.

- Removed `WIDE_VNI_VALUES` compile-time conditional and cmake
  option.

- Updated ovs submodule reference

+++ ovs-p4rt/include/ovsp4rt/ovs-p4rt.h
@@ -86,7 +86,7 @@ struct tunnel_info {
   struct p4_ipaddr local_ip;
   struct p4_ipaddr remote_ip;
   uint16_t dst_port;  // bit<16>
-  uint16_t vni;       // bit<24>
+  uint32_t vni;       // bit<24>
   struct port_vlan_info vlan_info;
   uint8_t bridge_id;    // bit<8>
   uint8_t tunnel_type;  // ovs_tunnel_type
